Added the backgroundScript. Converted popup.tsx to popup.jsx temporarily for testing. Made popup work (icons not appearing).
